Forward-proof a bit
Make the listening work with newer versions of psycopg too.  The main
change is that cursor objects no longer have a fileno method, so we
have to use connections and cursors depending on psycopg version.
This commit accidentially also fixes a bug where we would get a
backtrace if the database closed the connection on us.